third_party/blink/web_tests/fast/mediastream/enabled-expected.txt

Tests that navigator.webkitGetUserMedia is present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS typeof navigator.webkitGetUserMedia == 'function' is true
PASS hasGetUserMediaProperty() is true
PASS 'webkitGetUserMedia' in navigator is true
PASS navigator.hasOwnProperty('webkitGetUserMedia') is false
PASS successfullyParsed is true

TEST COMPLETE
